+++ packages/core/src/builder/buildProject.ts
@@ -3,6 +3,7 @@ import { SCHEMA_VERSION } from "../config";
 import { getNextRevision } from "./revision";
 import { buildDatafile, getCustomDatafile, buildScopedDatafile } from "./buildDatafile";
 import { Dependencies } from "../dependencies";
+import { Plugin } from "../cli";
 
 export interface BuildCLIOptions {
   revision?: string;
@@ -12,6 +13,7 @@ export interface BuildCLIOptions {
   feature?: string;
   print?: boolean;
   pretty?: boolean;
+  stateFiles?: boolean; // --no-state-files in CLI
 }
 
 export async function buildProject(deps: Dependencies, cliOptions: BuildCLIOptions = {}) {
@@ -99,12 +101,55 @@ export async function buildProject(deps: Dependencies, cliOptions: BuildCLIOptio
       }
     }
 
-    // write state for environment
-    await datasource.writeState(environment, existingState);
+    if (typeof cliOptions.stateFiles === "undefined" || cliOptions.stateFiles) {
+      // write state for environment
+      await datasource.writeState(environment, existingState);
 
-    // write revision
-    await datasource.writeRevision(nextRevision);
+      // write revision
+      await datasource.writeRevision(nextRevision);
+    }
   }
 
   console.log("\nLatest revision:", nextRevision);
 }
+
+export const buildPlugin: Plugin = {
+  command: "build",
+  handler: async function ({ rootDirectoryPath, projectConfig, datasource, parsed }) {
+    await buildProject(
+      {
+        rootDirectoryPath,
+        projectConfig,
+        datasource,
+        options: parsed,
+      },
+      parsed as BuildCLIOptions,
+    );
+  },
+  examples: [
+    {
+      command: "build",
+      description: "build datafiles for all environments and tags",
+    },
+    {
+      command: "build --revision=123",
+      description: "build datafiles starting with revision value as 123",
+    },
+    {
+      command: "build --environment=production",
+      description: "build datafiles for production environment",
+    },
+    {
+      command: "build --print --environment=production --feature=featureKey",
+      description: "print datafile for a feature in production environment",
+    },
+    {
+      command: "build --print --environment=production --pretty",
+      description: "print datafile with pretty print",
+    },
+    {
+      command: "build --no-state-files",
+      description: "build datafiles without writing state files",
+    },
+  ],
+};