+++ packages/core/src/config/projectConfig.ts
@@ -4,6 +4,7 @@ import { BucketBy, Context, Tag } from "@featurevisor/types";
 
 import { Parser, parsers } from "./parsers";
 import { FilesystemAdapter } from "../datasource/filesystemAdapter";
+import type { Plugin } from "../cli";
 
 export const FEATURES_DIRECTORY_NAME = "features";
 export const SEGMENTS_DIRECTORY_NAME = "segments";
@@ -49,22 +50,16 @@ export interface ProjectConfig {
   parser: Parser;
   prettyState: boolean;
   prettyDatafile: boolean;
+  stringify: boolean;
   siteExportDirectoryPath: string;
+  enforceCatchAllRule?: boolean;
   adapter: any; // @TODO: type this properly later
+  plugins: Plugin[];
 }
 
 // rootDirectoryPath: path to the root directory of the project (without ending with a slash)
 export function getProjectConfig(rootDirectoryPath: string): ProjectConfig {
   const baseConfig: ProjectConfig = {
-    featuresDirectoryPath: path.join(rootDirectoryPath, FEATURES_DIRECTORY_NAME),
-    segmentsDirectoryPath: path.join(rootDirectoryPath, SEGMENTS_DIRECTORY_NAME),
-    attributesDirectoryPath: path.join(rootDirectoryPath, ATTRIBUTES_DIRECTORY_NAME),
-    groupsDirectoryPath: path.join(rootDirectoryPath, GROUPS_DIRECTORY_NAME),
-    testsDirectoryPath: path.join(rootDirectoryPath, TESTS_DIRECTORY_NAME),
-
-    stateDirectoryPath: path.join(rootDirectoryPath, STATE_DIRECTORY_NAME),
-    outputDirectoryPath: path.join(rootDirectoryPath, OUTPUT_DIRECTORY_NAME),
-
     environments: DEFAULT_ENVIRONMENTS,
     tags: DEFAULT_TAGS,
     scopes: [],
@@ -74,10 +69,21 @@ export function getProjectConfig(rootDirectoryPath: string): ProjectConfig {
 
     prettyState: DEFAULT_PRETTY_STATE,
     prettyDatafile: DEFAULT_PRETTY_DATAFILE,
-
-    siteExportDirectoryPath: path.join(rootDirectoryPath, SITE_EXPORT_DIRECTORY_NAME),
+    stringify: true,
 
     adapter: FilesystemAdapter,
+
+    featuresDirectoryPath: path.join(rootDirectoryPath, FEATURES_DIRECTORY_NAME),
+    segmentsDirectoryPath: path.join(rootDirectoryPath, SEGMENTS_DIRECTORY_NAME),
+    attributesDirectoryPath: path.join(rootDirectoryPath, ATTRIBUTES_DIRECTORY_NAME),
+    groupsDirectoryPath: path.join(rootDirectoryPath, GROUPS_DIRECTORY_NAME),
+    testsDirectoryPath: path.join(rootDirectoryPath, TESTS_DIRECTORY_NAME),
+    stateDirectoryPath: path.join(rootDirectoryPath, STATE_DIRECTORY_NAME),
+    outputDirectoryPath: path.join(rootDirectoryPath, OUTPUT_DIRECTORY_NAME),
+    siteExportDirectoryPath: path.join(rootDirectoryPath, SITE_EXPORT_DIRECTORY_NAME),
+
+    enforceCatchAllRule: false,
+    plugins: [],
   };
 
   const configModulePath = path.join(rootDirectoryPath, CONFIG_MODULE_NAME);
@@ -86,7 +92,8 @@ export function getProjectConfig(rootDirectoryPath: string): ProjectConfig {
   const mergedConfig = {};
 
   Object.keys(baseConfig).forEach((key) => {
-    mergedConfig[key] = customConfig[key] || baseConfig[key];
+    mergedConfig[key] =
+      typeof customConfig[key] !== "undefined" ? customConfig[key] : baseConfig[key];
 
     if (key.endsWith("Path") && mergedConfig[key].indexOf(ROOT_DIR_PLACEHOLDER) !== -1) {
       mergedConfig[key] = mergedConfig[key].replace(ROOT_DIR_PLACEHOLDER, rootDirectoryPath);
@@ -106,3 +113,60 @@ export function getProjectConfig(rootDirectoryPath: string): ProjectConfig {
 
   return finalConfig as ProjectConfig;
 }
+
+export interface ShowProjectConfigOptions {
+  print?: boolean;
+  pretty?: boolean;
+}
+
+export function showProjectConfig(
+  projectConfig: ProjectConfig,
+  options: ShowProjectConfigOptions = {},
+) {
+  if (options.print) {
+    console.log(
+      options.pretty ? JSON.stringify(projectConfig, null, 2) : JSON.stringify(projectConfig),
+    );
+
+    return;
+  }
+
+  console.log("\nProject configuration:\n");
+
+  const keys = Object.keys(projectConfig);
+  const longestKeyLength = keys.reduce((acc, key) => (key.length > acc ? key.length : acc), 0);
+  const ignoreKeys = ["adapter", "parser"];
+
+  for (const key of keys) {
+    if (ignoreKeys.indexOf(key) !== -1) {
+      continue;
+    }
+
+    console.log(`  - ${key.padEnd(longestKeyLength, " ")}: ${projectConfig[key]}`);
+  }
+}
+
+export const configPlugin: Plugin = {
+  command: "config",
+  handler: async ({ rootDirectoryPath, parsed }) => {
+    const projectConfig = getProjectConfig(rootDirectoryPath);
+    showProjectConfig(projectConfig, {
+      print: parsed.print,
+      pretty: parsed.pretty,
+    });
+  },
+  examples: [
+    {
+      command: "config",
+      description: "show the project configuration",
+    },
+    {
+      command: "config --print",
+      description: "show the project configuration as JSON",
+    },
+    {
+      command: "config --print --pretty",
+      description: "show the project configuration (as pretty JSON)",
+    },
+  ],
+};